+++ bin/commands/pack_command.dart
@@ -36,7 +36,7 @@ class PackCommand extends Command {
       'clean',
       negatable: true,
       defaultsTo: true,
-      help: 'Remove ${config.dist.path} output after command succeed.',
+      help: 'Remove ${config.stage.path} output after command succeed.',
     );
   }