+++ bin/data/script_language.dart
@@ -1,0 +1,1 @@
+enum ScriptLanguage { redscript, cet, red4ext }